+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.0.1] - 2026-03-29
+
+- Added a GitHub release version checker that periodically looks for newer plugin versions.
+- Added join-time update notifications for server operators with a direct download link when an update is available.
+
 ## [1.0.0] - 2026-03-27
 
 - First release of Majestic-Spawners.